cover and grouping
cover now exists and has their respective points of interest. also points of interest now has groupings of objects. debug gizmos are added to show enemy groups, freindly groups and cover within the awareness range of the unit.

+++ MartinJonesFYP/Assets/bullet.cs
@@ -5,7 +5,7 @@ using UnityEngine;
 public class bullet : MonoBehaviour
 {
     public Vector3 m_direction;
-    public int m_damage;
+    public float m_damage;
     public bool m_isPlayerOwned;
 
     public void cleanup()
